refactor: move checksum options into their own file
