+++ packages/storage/Sources/GoogleCloudStorage/UploadOptions.swift
@@ -16,81 +16,6 @@ import Crypto
 import Foundation
 import GoogleCloudWkt
 
-/// Strategy for data integrity validation.
-public enum ChecksumValidation: Sendable {
-  /// Do not perform client-side checksum validation.
-  case none
-
-  /// Automatically calculate and validate CRC32C (recommended).
-  case crc32c
-
-  /// Automatically calculate and validate MD5.
-  case md5
-}
-
-/// Configuration options for upload checksum validation.
-public struct ChecksumOptions: Sendable, Hashable {
-  /// Checksum mode / value for CRC32C.
-  public var crc32c: ChecksumValue?
-
-  /// Checksum mode / value for MD5.
-  public var md5: ChecksumValue?
-
-  /// Specifies how a checksum should be provided for upload validation.
-  public enum ChecksumValue: Sendable, Hashable, ExpressibleByStringLiteral,
-    ExpressibleByIntegerLiteral
-  {
-    /// Automatically calculate the checksum on-the-fly during upload streaming.
-    case auto
-
-    /// Use a pre-computed checksum value (e.g., Base64 encoded string).
-    case value(String)
-
-    /// Creates a `ChecksumValue` from a string literal containing a pre-computed checksum.
-    public init(stringLiteral value: String) {
-      self = .value(value)
-    }
-
-    /// Creates a `ChecksumValue` from a 32-bit unsigned integer CRC32C checksum value.
-    public init(_ intValue: UInt32) {
-      let bigEndian = intValue.bigEndian
-      let base64 = withUnsafeBytes(of: bigEndian) { Data($0).base64EncodedString() }
-      self = .value(base64)
-    }
-
-    /// Creates a `ChecksumValue` from an integer literal containing a CRC32C checksum value.
-    public init(integerLiteral value: UInt64) {
-      self.init(UInt32(truncatingIfNeeded: value))
-    }
-  }
-
-  /// Creates a new `ChecksumOptions` configuration for validating uploads with Google Cloud Storage.
-  ///
-  /// You can configure automatic on-the-fly calculation (`.auto`), pre-computed values (e.g., `.value("...")` or a string literal `"..."`),
-  /// or enable multiple checksum types simultaneously in a single upload request. `crc32c` is the default and recommended checksum option,
-  /// as it provides better computational performance compared to MD5.
-  public init(crc32c: ChecksumValue? = .auto, md5: ChecksumValue? = nil) {
-    self.crc32c = crc32c
-    self.md5 = md5
-  }
-
-  /// Default options: Automatically calculate CRC32C on-the-fly.
-  public static var `default`: ChecksumOptions {
-    ChecksumOptions(crc32c: .auto, md5: nil)
-  }
-
-  /// No checksum validation.
-  public static var none: ChecksumOptions {
-    ChecksumOptions(crc32c: nil, md5: nil)
-  }
-
-  public var hasUserProvidedChecksum: Bool {
-    if case .value = crc32c { return true }
-    if case .value = md5 { return true }
-    return false
-  }
-}
-
 /// Errors thrown when validating or creating a `CustomerEncryptionKeyOptions`.
 public enum CustomerEncryptionKeyError: Error, Sendable, Equatable,
   CustomStringConvertible
